cvs_slm_assistant/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
streamlit>=1.28.0
requests>=2.31.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pypdf>=3.0.0
docx2txt>=0.8
pandas>=2.0.0
